Offload some PaStmtSrc class code into a mixin.
