+++ indra_db/schemas/mixins.py
@@ -1,5 +1,7 @@
 import logging
 from psycopg2.errors import DuplicateTable
+from sqlalchemy import inspect, Column, BigInteger
+from sqlalchemy.exc import NoSuchTableError
 
 logger = logging.getLogger(__name__)
 
@@ -116,6 +118,40 @@ class ReadonlyTable(IndraDBTable):
         cursor = conn.cursor()
         cursor.execute(sql)
         conn.commit()
+        return
+
+
+class SpecialColumnTable(ReadonlyTable):
+
+    @classmethod
+    def create(cls, db, commit=True):
+
+        # Create the materialized view.
+        cls.__definition__ = cls.definition(db)
+        sql = super(cls, cls).create(db, commit)
+        cls.loaded = True
+        return sql
+
+    @classmethod
+    def load_cols(cls, engine):
+        if cls.loaded:
+            return
+
+        try:
+            schema = cls.__table_args__.get('schema', 'public')
+            cols = inspect(engine).get_columns(cls.__tablename__,
+                                               schema=schema)
+        except NoSuchTableError:
+            return
+
+        existing_cols = {col.name for col in cls.__table__.columns}
+        for col in cols:
+            if col['name'] in existing_cols:
+                continue
+
+            setattr(cls, col['name'], Column(BigInteger))
+
+        cls.loaded = True
         return
